refactor : exception.player 4개 이상 input 제한 익셉션 분리
