+++ src/main/java/baseball/exception/player/InputSizeException.java
@@ -1,0 +1,9 @@
+package baseball.exception.player;
+
+public class InputSizeException extends IllegalArgumentException {
+    private static final String INPUT_SIZE_MESSAGE = "3개 이상의 공이 입력될 수 없습니다.";
+
+    public InputSizeException() {
+        super(INPUT_SIZE_MESSAGE);
+    }
+}
